pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glm-ocr-example-eval"
version = "0.1.0"
description = "Starter scorer for GLM-OCR-Swift example parity and quality evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "markdown-it-py>=3.0.0",
  "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
example-eval = "example_eval.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
